fix(seo): put the find-funders hero in the visible no-JS HTML
QA on the previous commit found that while the find-funders landing page now
server-renders, its HTML arrives as a React streaming payload: a
`<div hidden id=\"S:n\">` chunk revealed only by client-side script, with a
loading fallback as the visible page. A reader without JavaScript still saw a
headingless spinner shell.

Streaming instrumentation across six builds showed why no in-segment fix can
hold: every route renders dynamically (the root layout reads request headers
for whitelabel detection), and with a loading.tsx anywhere above a page, Next
always defers the page behind a Suspense boundary. Which ancestor fallback
ends up visible is decided by module-chunk timing and flips between builds —
a route-group layout, force-static (Next serializes the fallback + hidden
chunk even into prerendered HTML), and an import-free meaningful fallback all
failed against the built output. The only region React guarantees into the
initially visible HTML is the root layout's own output, outside every
boundary.

So the hero for no-JS readers renders there:

- the proxy tags /nonprofits/find-funders requests with x-pathname (that one
  path only);
- the root layout renders a <noscript> static hero — h1, lead copy, example
  prompts — when the tag matches. JS-executing visitors never display it; they
  get the interactive hero when hydration reveals the page, so nothing is
  duplicated in either mode;
- hero copy moves to lib/hero-content, shared by the interactive hero and the
  noscript replica so the two cannot drift, and the interactive hero splits
  into its own component file.

Verified against the built output: the h1 now sits outside any hidden element
in the served HTML, hydration still yields exactly one visible h1, and the
/communities and /seeds fixes are unaffected. Tests pin the noscript hero's
content, its h1 parity with the page, the proxy header contract, and the
per-subroute loading files.
